requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
yagmail==0.15.293
